refactor: move `stop_extension_background_work` to extension manager
