+++ lnbits/extension_manager.py
@@ -1,5 +1,6 @@
 import asyncio
 import hashlib
+import importlib
 import json
 import os
 import shutil
@@ -770,6 +771,68 @@ class ExtensionDetailsRequest(BaseModel):
     ext_id: str
     source_repo: str
     version: str
+
+
+async def stop_extension_background_work(
+    ext_id: str, user: str, access_token: Optional[str] = None
+):
+    """
+    Stop background work for extension (like asyncio.Tasks, WebSockets, etc).
+    Extensions SHOULD expose a `api_stop()` function and/or a DELETE enpoint
+    at the root level of their API.
+    """
+    stopped = await _stop_extension_background_work(ext_id)
+
+    if not stopped:
+        # fallback to REST API call
+        await _stop_extension_background_work_via_api(ext_id, user, access_token)
+
+
+async def _stop_extension_background_work(ext_id) -> bool:
+    upgrade_hash = settings.extension_upgrade_hash(ext_id) or ""
+    ext = Extension(ext_id, True, False, upgrade_hash=upgrade_hash)
+
+    try:
+        logger.info(f"Stopping background work for extension '{ext.module_name}'.")
+        old_module = importlib.import_module(ext.module_name)
+
+        # Extensions must expose an `{ext_id}_stop()` function at the module level
+        # The `api_stop()` function is for backwards compatibility (will be deprecated)
+        stop_fns = [f"{ext_id}_stop", "api_stop"]
+        stop_fn_name = next((fn for fn in stop_fns if hasattr(old_module, fn)), None)
+        assert stop_fn_name, "No stop function found for '{ext.module_name}'"
+
+        stop_fn = getattr(old_module, stop_fn_name)
+        if stop_fn:
+            await stop_fn()
+
+        logger.info(f"Stopped background work for extension '{ext.module_name}'.")
+    except Exception as ex:
+        logger.warning(f"Failed to stop background work for '{ext.module_name}'.")
+        logger.warning(ex)
+        return False
+
+    return True
+
+
+async def _stop_extension_background_work_via_api(ext_id, user, access_token):
+    logger.info(
+        f"Stopping background work for extension '{ext_id}' using the REST API."
+    )
+    async with httpx.AsyncClient() as client:
+        try:
+            url = f"http://{settings.host}:{settings.port}/{ext_id}/api/v1?usr={user}"
+            headers = (
+                {"Authorization": "Bearer " + access_token} if access_token else None
+            )
+            resp = await client.delete(url=url, headers=headers)
+            resp.raise_for_status()
+            logger.info(f"Stopped background work for extension '{ext_id}'.")
+        except Exception as ex:
+            logger.warning(
+                f"Failed to stop background work for '{ext_id}' using the REST API."
+            )
+            logger.warning(ex)
 
 
 def get_valid_extensions(include_deactivated: Optional[bool] = True) -> List[Extension]:
